homefinderbackend/requirement.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==3.7.1
attrs==23.2.0
bcrypt==4.1.2
certifi==2024.2.2
cffi==1.16.0
click==8.1.7
colorama==0.4.6
coverage==7.4.1
cryptography==41.0.7
distlib==0.3.7
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
fastapi==0.104.1
filelock==3.13.1
fuzzywuzzy==0.18.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
iniconfig==2.0.0
joblib==1.3.2
Levenshtein==0.23.0
Mako==1.3.0
MarkupSafe==2.1.3
nltk==3.8.1
outcome==1.3.0.post0
packaging==23.2
passlib==1.7.4
platformdirs==3.11.0
pluggy==1.4.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.5.1
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
pytest==7.4.4
pytest-cov==4.1.0
pytest-trio==0.8.0
python-jose==3.3.0
python-Levenshtein==0.23.0
python-multipart==0.0.6
rapidfuzz==3.6.1
regex==2023.12.25
rsa==4.9
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.25
starlette==0.27.0
textblob==0.17.1
tqdm==4.66.1
trio==0.24.0
typing_extensions==4.8.0
uvicorn==0.25.0
virtualenv==20.24.6
